migrate Clam AV action tests
